requirements.txt
#run: pip install -r requirements.txt
replicate
python-dotenv
qrcode
google-api-python-client
google-cloud-storage